Remove unused armor JSON files (Kweebec, Trooper, Trork Warrior), added functional repair bench, added game mode npc selection

+++ src/main/java/dev/hytalemodding/origins/interaction/FishingInteraction.java
@@ -341,7 +341,7 @@ public class FishingInteraction extends SimpleInstantInteraction {
             updated = current.withMetadata(FishingMetaData.KEYED_CODEC, metaData);
         }
 
-        hotbar.replaceItemStackInSlot((short) slot, current, updated);
+        hotbar.replaceItemStackInSlot(slot, current, updated);
     }
 
     public static void playCastSound(PlayerRef playerRef) {
